requirements.txt
certifi==2018.8.24
chardet==3.0.4
Django==2.1.1
djangorestframework==3.9.0
idna==2.6
python-dateutil==2.7.3
pytz==2018.5
requests==2.18.4
Robinhood==1.0.1
selenium==3.14.0
six==1.11.0
urllib3==1.22
